Guard Equality
In a ActiveRecord before_validation callback, I was returning an instance of VideoInfo to indicate success. I learned, however, that before_validation callbacks compares the method's result to false in order to continue or abort the callback chain. When comparing an instance of VideoInfo against false it will throw a NoMethodError by trying to call "url" on false:FalseClass... Maybe its a good idea to guard the comparison.